1. Merge statementListener to sqlEngineListener
