+++ externals/kyuubi-spark-sql-engine/src/main/scala/org/apache/kyuubi/engine/spark/SparkSQLEngine.scala
@@ -27,7 +27,6 @@ import org.apache.spark.sql.SparkSession
 import org.apache.kyuubi.{Logging, Utils}
 import org.apache.kyuubi.config.KyuubiConf
 import org.apache.kyuubi.engine.spark.SparkSQLEngine.countDownLatch
-import org.apache.kyuubi.engine.spark.monitor.listener.KyuubiStatementListener
 import org.apache.kyuubi.ha.HighAvailabilityConf._
 import org.apache.kyuubi.ha.client.{EngineServiceDiscovery, RetryPolicies, ServiceDiscovery}
 import org.apache.kyuubi.service.{Serverable, Service}
@@ -48,7 +47,6 @@ private[spark] final class SparkSQLEngine(name: String, spark: SparkSession)
   override def initialize(conf: KyuubiConf): Unit = {
     val listener = new SparkSQLEngineListener(this)
     spark.sparkContext.addSparkListener(listener)
-    spark.sparkContext.addSparkListener(new KyuubiStatementListener)
     super.initialize(conf)
   }
